ENHANCEMENT: Tidying up namespacing
