+++ src/Rules/SingleSymbolRule.php
@@ -1,5 +1,8 @@
 <?php
-namespace Veer66\PhlongTaIam;
+namespace Veer66\PhlongTaIam\Rules;
+
+use Veer66\PhlongTaIam\Acceptor\SingleSymbolAcceptor;
+
 
 class SingleSymbolRule
 {
