+++ src/Rules/WordRule.php
@@ -1,5 +1,7 @@
 <?php
-namespace Veer66\PhlongTaIam;
+namespace Veer66\PhlongTaIam\Rules;
+
+use Veer66\PhlongTaIam\Acceptor\WordRuleAcceptor;
 
 class WordRule
 {
